+++ bin/polymer_app.dart
@@ -11,13 +11,22 @@ library polymer.bin.new_element;
 
 import 'dart:io';
 import 'package:args/args.dart';
-import 'package:polymer_app/templates.dart';
 import 'package:ansicolor/ansicolor.dart';
 import 'package:polymer_app/utils.dart';
+import "package:polymer_app/polymer_app_manager.dart";
+import "package:polymer_app/polymer_app_services.dart";
+import "package:polymer_app/polymer_app_models.dart";
+import "package:polymer_app/polymer_app_behaviors.dart";
+import "package:polymer_app/polymer_app_elements.dart";
+import "package:polymer_app/polymer_app_routes.dart";
 
 final AnsiPen green = new AnsiPen()..green(bold: true);
-
 final AnsiPen white = new AnsiPen()..white(bold: true);
+
+PolymerAppManager manager;
+Directory outputFolder;
+File config;
+String outputFolderPath;
 
 String getAppName(ArgResults results, ArgParser parser) {
   String appName;
@@ -42,159 +51,78 @@ Directory getDirectory(String appName, ArgParser parser) {
   return dir;
 }
 
-Directory createWebDirectory(Directory root, String appName) {
-  print("");
-  Directory webDirectory =
-      createDirectory("${root.resolveSymbolicLinksSync()}/web");
-
-  File indexHtmlFile =
-      createFile("${webDirectory.resolveSymbolicLinksSync()}/index.html");
-  indexHtmlFile.createSync(recursive: true);
-  indexHtmlFile.writeAsStringSync(indexHtmlContent(appName));
-
-  File indexDartFile =
-      createFile("${webDirectory.resolveSymbolicLinksSync()}/index.dart");
-  indexDartFile.createSync(recursive: true);
-  indexDartFile.writeAsStringSync(indexDartContent(appName));
-
-  return webDirectory;
-}
-
-Directory createModelsDirectory(Directory root, String appName) {
-  Directory modelsDirectory =
-      createDirectory("${root.resolveSymbolicLinksSync()}/models");
-  File file =
-      createFile("${modelsDirectory.resolveSymbolicLinksSync()}/models.dart");
-  file.writeAsStringSync(modelsLibContent(appName));
-  return modelsDirectory;
-}
-
-Directory createBehaviorsDirectory(Directory root, String appName) {
-  Directory behaviorsDirectory =
-      createDirectory("${root.resolveSymbolicLinksSync()}/behaviors");
-  File file = createFile(
-      "${behaviorsDirectory.resolveSymbolicLinksSync()}/behaviors.dart");
-  file.writeAsStringSync(behaviorsLibContent(appName));
-  return behaviorsDirectory;
-}
-
-Directory createServicesDirectory(Directory root, String appName) {
-  Directory servicesDirectory =
-      createDirectory("${root.resolveSymbolicLinksSync()}/services");
-  File file = createFile(
-      "${servicesDirectory.resolveSymbolicLinksSync()}/services.dart");
-  file.writeAsStringSync(servicesLibContent(appName));
-  return servicesDirectory;
-}
-
-Directory createElementsDirectory(Directory root, String appName) {
-  Directory elementsDirectory =
-      createDirectory("${root.resolveSymbolicLinksSync()}/elements");
-  File file = createFile(
-      "${elementsDirectory.resolveSymbolicLinksSync()}/elements.dart");
-  file.writeAsStringSync(elementsLibContent(appName));
-
-  Directory dir = createDirectory(
-      "${elementsDirectory.resolveSymbolicLinksSync()}/root_element");
-
-  File fileDart =
-      createFile("${dir.resolveSymbolicLinksSync()}/root_element.dart");
-  fileDart.writeAsStringSync(rootElementDartContent(appName));
-
-  File fileHtml =
-      createFile("${dir.resolveSymbolicLinksSync()}/root_element.html");
-  fileHtml.writeAsStringSync(rootElementHtmlContent());
-
-  File fileCss =
-      createFile("${dir.resolveSymbolicLinksSync()}/root_element.css");
-
-  fileCss.writeAsStringSync(rootElementCssContent());
-  return elementsDirectory;
-}
-
-Directory createLibDirectory(Directory root, String appName) {
-  print("");
-  Directory libDirectory =
-      createDirectory("${root.resolveSymbolicLinksSync()}/lib");
-
-  createModelsDirectory(libDirectory, appName);
-  createServicesDirectory(libDirectory, appName);
-  createBehaviorsDirectory(libDirectory, appName);
-  createElementsDirectory(libDirectory, appName);
-
-  File appNameFile = createFile(
-      "${libDirectory.resolveSymbolicLinksSync()}/${toSnakeCase(appName)}.dart");
-  appNameFile.createSync(recursive: true);
-  appNameFile.writeAsStringSync(appLibraryContent(appName));
-
-  return libDirectory;
-}
-
-createPubspec(Directory directory, String appName) {
-  print("");
-  File pubspecFile =
-      createFile("${directory.resolveSymbolicLinksSync()}/pubspec.yaml");
-  pubspecFile.createSync(recursive: true);
-  pubspecFile.writeAsStringSync(pubspecContent(appName));
-}
-
-createNewApplication(ArgResults results, ArgParser parser, String appName) {
+createNewApplication(ArgResults results) {
+  String appName = results.rest[2];
   print("Creating '${green(appName)}' application");
-  Directory directory = getDirectory(appName, parser);
-
-  createPubspec(directory, appName);
-  createLibDirectory(directory, appName);
-  createWebDirectory(directory, appName);
+  writeInFile("${outputFolder.resolveSymbolicLinksSync()}/polymer_app.json",
+      getDefaultJsonConfig(appName));
+  manager.createApplication();
 }
 
-createNewService(ArgResults results, ArgParser parser, String elementName) {
-  Directory dir = new Directory(".");
-  String library =
-      toSnakeCase(dir?.resolveSymbolicLinksSync()?.split("/")?.last);
-  createService(
-      elementName, dir.resolveSymbolicLinksSync() + "/lib/services/", library);
+createNewService(String name) {
+  print("Creating '${green(name)}' service");
+
+  ServicesManager services =
+      manager?.services ?? new ServicesManager(name, outputFolderPath);
+
+  services.createService(name);
+  if (manager != null) {
+    services.addToLibrary(name);
+  }
+
 }
 
-createNewModel(ArgResults results, ArgParser parser, String elementName) {
-  Directory dir = new Directory(".");
-  String library =
-      toSnakeCase(dir?.resolveSymbolicLinksSync()?.split("/")?.last);
-  createModel(
-      elementName, dir.resolveSymbolicLinksSync() + "/lib/models/", library);
+createNewModel(String name) {
+  print("Creating '${green(name)}' model");
+  ModelsManager models =
+      manager?.models ?? new ModelsManager(name, outputFolderPath);
+
+  models.createModel(name);
+  if (manager != null) {
+    models.addToLibrary(name);
+  }
 }
 
-createNewElement(ArgResults results, ArgParser parser, String elementName) {
-  if (toLispCase(elementName).split("-")?.length < 2) {
+createNewElement(String name) {
+  if (toLispCase(name).split("-")?.length < 2) {
     print("Bad element name, should be 'polymer-element'");
     exit(1);
   }
-  Directory dir = new Directory(".");
-  String library =
-      toSnakeCase(dir?.resolveSymbolicLinksSync()?.split("/")?.last);
-  createPolymerElement(
-      elementName, dir.resolveSymbolicLinksSync() + "/lib/elements/", library);
+
+  ElementsManager elements =
+      manager?.elements ?? new ElementsManager(name, outputFolderPath);
+
+  print("Creating '${green(name)}' element");
+  elements.createElement(name);
+  if (manager != null) {
+    elements.addToLibrary(name);
+  }
 }
 
-createNewBehavior(ArgResults results, ArgParser parser, String elementName) {
-  Directory dir = new Directory(".");
-  String library =
-      toSnakeCase(dir?.resolveSymbolicLinksSync()?.split("/")?.last);
-  createPolymerBehavior(
-      elementName, dir.resolveSymbolicLinksSync() + "/lib/behaviors/", library);
+createNewBehavior(String name) {
+  BehaviorsManager behaviors = manager?.behaviors ?? new BehaviorsManager(name, outputFolderPath);
+  print(outputFolderPath);
+
+  print("Creating '${green(name)}' behavior");
+  behaviors.createBehavior(name);
+  if (manager != null) {
+    behaviors.addToLibrary(name);
+  }
 }
 
-createNewRoute(
-    ArgResults results, ArgParser parser, String routeName, String path) {
-  Directory dir = new Directory(".");
-  String library =
-      toSnakeCase(dir?.resolveSymbolicLinksSync()?.split("/")?.last);
-  createRouteElement(
-      routeName, dir.resolveSymbolicLinksSync() + "/lib/elements/", library);
+createNewRoute(String routeName, String path) {
+  print("Creating '${green(routeName)}' route");
 
+  RoutesManager routes = manager?.routes ?? new RoutesManager("$routeName-route", outputFolderPath);
+
+  routes.createRoute("$routeName-route");
   print(white("\nImport and add page to your root-element\n"));
-  print(green(
-          "import 'package:${toSnakeCase(library)}/elements/route/${toSnakeCase(routeName)}_route/${toSnakeCase(routeName)}_route.dart';") +
-      "\n");
+  if (manager != null) {
+    routes.addToLibrary("$routeName-route");
+    print(green(
+        "import 'package:${toSnakeCase(manager.name)}/routes/routes.dart';") +
+        "\n");
+  }
   print("List<Page> _pages = [");
   print("...");
   print(green(
@@ -206,37 +134,70 @@ createNewRoute(
 void main(List<String> args) {
   ArgParser parser = new ArgParser(allowTrailingOptions: true);
 
+  parser.addOption('output-folder', abbr: "o", defaultsTo: "./");
   parser.addFlag('help', abbr: 'h');
 
-  ArgResults results = parser.parse(args);
-
-  if (results.rest.length == 3 && results.rest[0] == "new") {
-    if (results.rest[1] == "app") {
-      return createNewApplication(results, parser, results.rest[2]);
-    } else if (results.rest[1] == "element") {
-      return createNewElement(results, parser, results.rest[2]);
-    } else if (results.rest[1] == "behavior") {
-      return createNewBehavior(results, parser, results.rest[2]);
-    } else if (results.rest[1] == "model") {
-      return createNewModel(results, parser, results.rest[2]);
-    } else if (results.rest[1] == "service") {
-      return createNewService(results, parser, results.rest[2]);
+  try {
+    ArgResults results = parser.parse(args);
+    outputFolderPath = results["output-folder"];
+    outputFolder = new Directory(outputFolderPath);
+    if (!outputFolder.existsSync()) {
+      outputFolder.createSync(recursive: true);
     }
-  } else if (results.rest.length == 4 &&
-      results.rest[0] == "new" &&
-      results.rest[1] == "route") {
-    return createNewRoute(results, parser, results.rest[2], results.rest[3]);
+    config = new File("$outputFolderPath/polymer_app.json");
+
+    if (config.existsSync()) {
+      manager = new PolymerAppManager(config.resolveSymbolicLinksSync(),
+          outputFolder.resolveSymbolicLinksSync());
+    }
+
+    if (isCommandNew(results.rest, "app")) {
+      if (manager == null) {
+        manager = new PolymerAppManager.fromJson(
+            getDefaultJsonConfig(results.rest[2]), outputFolder.resolveSymbolicLinksSync());
+      }
+      return createNewApplication(results);
+    } else if (isCommandNew(results.rest)) {
+      if (results.rest[1] == "element") {
+        return createNewElement(results.rest[2]);
+      } else if (results.rest[1] == "behavior") {
+        return createNewBehavior(results.rest[2]);
+      } else if (results.rest[1] == "model") {
+        return createNewModel(results.rest[2]);
+      } else if (results.rest[1] == "service") {
+        return createNewService(results.rest[2]);
+      } else if (results.rest.length == 4 && results.rest[1] == "route") {
+        return createNewRoute(results.rest[2], results.rest[3]);
+      } else {
+        usage(parser);
+      }
+    } else {
+      usage(parser);
+    }
+  } catch (e) {
+    print(e);
+    usage(parser);
   }
-  usage(parser);
 }
 
 void usage(ArgParser parser) {
   print('polymer_app \n'
-      ' - new app app_name\n'
-      ' - new element element-name\n'
-      ' - new model name\n'
-      ' - new behavior name\n'
-      ' - new service name\n'
-      ' - new route name path\n');
+      'new app app_name\n'
+      'new element element-name\n'
+      'new model name\n'
+      'new behavior name\n'
+      'new service name\n'
+      'new route name path\n');
   print(parser.usage);
 }
+
+String getDefaultJsonConfig(String appName, [String path = "lib"]) => '{'
+    '"name": "$appName",'
+    '"library_path": "$path",'
+    '"web_path": "web",'
+    '"elements_path": "elements",'
+    '"services_path": "services",'
+    '"behaviors_path": "behaviors",'
+    '"models_path":  "models",'
+    '"routes_path": "routes"'
+    '}';