+++ lib/utils.dart
@@ -5,10 +5,11 @@
 library polymer_app.utils;
 
 import 'dart:io';
-import 'package:polymer_app/templates.dart';
 import 'package:ansicolor/ansicolor.dart';
+import 'package:dart_style/dart_style.dart';
 
 final AnsiPen green = new AnsiPen()..green(bold: true);
+final AnsiPen white = new AnsiPen()..white(bold: true);
 
 Directory createDirectory(String path) {
   print("Creating '${green(path)}' directory.");
@@ -28,91 +29,27 @@ File createFile(String path) {
   return file;
 }
 
-createService(String name, String path, String appName) {
-  Directory dir = createDirectory("$path${toSnakeCase(name)}");
-
-  File services = new File("$path/services.dart");
-  services.writeAsString(services.readAsStringSync() +
-      "\n" +
-      "export '${toSnakeCase(name)}/${toSnakeCase(name)}.dart';\n");
-
-  File fileDart =
-  createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.dart");
-  fileDart.writeAsStringSync(factoryContent(name, appName));
-}
-
-createPolymerBehavior(String name, String path, String appName) {
-  Directory dir = createDirectory("$path${toSnakeCase(name)}");
-
-  File behaviors = new File("$path/behaviors.dart");
-  behaviors.writeAsString(behaviors.readAsStringSync() +
-      "\n" +
-      "export '${toSnakeCase(name)}/${toSnakeCase(name)}.dart';\n");
-
-  File fileDart =
-  createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.dart");
-  fileDart.writeAsStringSync(polymerBehaviorContent(name, appName));
-}
-
-createRouteElement(String name, String path, String appName) {
-  name = name + "-route";
-  Directory dir = createDirectory("${path}route/${toSnakeCase(name)}");
-
-  File elements = new File("$path/elements.dart");
-  elements.writeAsString(elements.readAsStringSync() +
-      "\n" +
-      "export 'route/${toSnakeCase(name)}/${toSnakeCase(name)}.dart';\n");
-
-  File fileDart =
-  createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.dart");
-  fileDart.writeAsStringSync(polymerElementDartRouteContent(name, appName));
-
-  File fileHtml =
-  createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.html");
-  fileHtml.writeAsStringSync(polymerElementHtmlContent(name));
-
-  File fileCss =
-  createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.css");
-  fileCss.writeAsStringSync(polymerElementCssContent());
-}
-
-createPolymerElement(String name, String path, String appName) {
-  Directory dir = createDirectory("$path${toSnakeCase(name)}");
-
-  File elements = new File("$path/elements.dart");
-  elements.writeAsString(elements.readAsStringSync() +
-      "\n" +
-      "export '${toSnakeCase(name)}/${toSnakeCase(name)}.dart';\n");
-
-  File fileDart =
-      createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.dart");
-  fileDart.writeAsStringSync(polymerElementDartContent(name, appName));
-
-  File fileHtml =
-      createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.html");
-  fileHtml.writeAsStringSync(polymerElementHtmlContent(name));
-
-  File fileCss =
-      createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.css");
-  fileCss.writeAsStringSync(polymerElementCssContent());
-}
-
-createModel(String name, String path, String appName) {
-  Directory dir = createDirectory("$path${toSnakeCase(name)}");
-
-  File models = new File("$path/models.dart");
-  models.writeAsString(models.readAsStringSync() +
-      "\n" +
-      "export '${toSnakeCase(name)}/${toSnakeCase(name)}.dart';\n");
-
-  File fileDart =
-  createFile("${dir.resolveSymbolicLinksSync()}/${toSnakeCase(name)}.dart");
-  fileDart.writeAsStringSync(modelContent(name, appName));
-}
-
-toSnakeCase(String name) => name?.replaceAll("-", "_").replaceAll(" ", "_")?.toLowerCase();
-toLispCase(String name) => name?.replaceAll("_", "-").replaceAll(" ", "-")?.toLowerCase();
+toSnakeCase(String name) =>
+    name?.replaceAll("-", "_").replaceAll(" ", "_")?.toLowerCase();
+toLispCase(String name) =>
+    name?.replaceAll("_", "-").replaceAll(" ", "-")?.toLowerCase();
 toCamelCase(String str) => toLispCase(str)
     ?.split('-')
     ?.map((e) => e[0].toUpperCase() + e.substring(1))
     .join('');
+
+DartFormatter _formatter = new DartFormatter();
+
+writeInDartFile(String path, String content) {
+  writeInFile(path, _formatter.format(content));
+}
+
+writeInFile(String path, String content) {
+  File fileDart = createFile(path);
+  fileDart.writeAsStringSync(content);
+}
+
+
+bool isCommandNew(commands, [String arg]) => commands?.length >= 2 &&
+    commands[0] == "new" &&
+    (arg == null || commands[1] == arg);