+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.0
+- Refactor Manager
+- Implement -output-folder option
+
+## 0.0.7
+- Implement 'polymer_app.json' config
+
 ## 0.0.6
 - Solid base of tools